requirements.txt
auth==0.5.3
Flask==2.2.2
matplotlib==2.2.2
numpy==1.21.6
pandas==1.3.5
pandas_datareader==0.10.0
pandas_ta==0.3.14b0
plotly==5.16.1
python_binance==1.0.19
PyYAML==5.4.1
service==0.6.0
yfinance==0.2.28
